Refactor container builder to use providers
